+++ app/[username]/[slug]/page.tsx
@@ -1,4 +1,5 @@
 import AlternativeAccordion from "@/components/alternativeAccordion";
+import Banner from "@/components/banner";
 import CgsDeepLink from "@/components/cgsDeepLink";
 import { getGame } from "@/lib/firebase/firestore";
 import { notFound } from "next/navigation";
@@ -12,11 +13,18 @@ export default async function Page({
   if (!game) return notFound();
   const cgsgg = "https://cgs.gg/?url=" + encodeURIComponent(game.autoUpdateUrl);
   return (
-    <section className="main-content">
-      <h2>Launch the CGS app to play {game.name}:</h2>
-      <CgsDeepLink game={game} />
-      <h2>Alternatively:</h2>
-      <AlternativeAccordion game={game} cgsgg={cgsgg} />
-    </section>
+    <>
+      <Banner
+        home={`/` + params.username + `/` + params.slug}
+        img={game.bannerImageUrl}
+        txt={game.name}
+      />
+      <main className="main-content">
+        <h2>Launch the CGS app to play {game.name}:</h2>
+        <CgsDeepLink game={game} />
+        <h2>Alternatively:</h2>
+        <AlternativeAccordion game={game} cgsgg={cgsgg} />
+      </main>
+    </>
   );
 }
