+++ app/page.tsx
@@ -1,6 +1,5 @@
 import Footer from "@/components/footer";
 import GamesDeck from "@/components/gamesDeck";
-import Header from "@/components/header";
 import { getAllGames } from "@/lib/firebase/firestore";
 import Link from "next/link";
 
@@ -9,27 +8,30 @@ export const dynamic = "force-dynamic";
 export default async function Page() {
   const allGames = await getAllGames();
   return (
-    <main className="main-container">
-      <Header title="CGS Games" />
-      <div className="main-content">
+    <>
+      <main className="main-content">
+        <h1 className="text-center">CGS Games</h1>
         <p>
           Welcome to the{" "}
-          <Link href="https://www.cardgamesimulator.com">
+          <Link href="https://www.cardgamesimulator.com" target="_blank">
             Card Game Simulator
           </Link>{" "}
           (CGS) Games website!
         </p>
         <p>
           You can{" "}
-          <Link href="https://github.com/finol-digital/Card-Game-Simulator/wiki/Crash-Course-into-Game-Development-with-CGS">
+          <Link
+            href="https://github.com/finol-digital/Card-Game-Simulator/wiki/Crash-Course-into-Game-Development-with-CGS"
+            target="_blank"
+          >
             create
           </Link>{" "}
           and <Link href="/upload">upload</Link> your own custom card games, or
           browse games uploaded by others:
         </p>
         <GamesDeck games={allGames} />
-      </div>
+      </main>
       <Footer />
-    </main>
+    </>
   );
 }
