+++ app/privacy/page.tsx
@@ -1,5 +1,3 @@
-import Footer from "@/components/footer";
-import Header from "@/components/header";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -70,10 +68,9 @@ word-break: break-word !important;
 `;
 export default function Page() {
   return (
-    <main className="main-container">
-      <Header title="CGS Games" />
+    <main>
+      <h1 className="text-center">CGS Games Privacy Policy</h1>
       <div dangerouslySetInnerHTML={{ __html: rawHTML }}></div>
-      <Footer />
     </main>
   );
 }
