+++ app/terms/page.tsx
@@ -1,5 +1,3 @@
-import Footer from "@/components/footer";
-import Header from "@/components/header";
 import { Metadata } from "next";
 
 export const metadata: Metadata = {
@@ -157,10 +155,9 @@ Calibri;color:#595959;mso-themecolor:text1;mso-themetint:166;"><strong><bdt clas
 
 export default function Page() {
   return (
-    <main className="main-container">
-      <Header title="CGS Games" />
+    <main>
+      <h1 className="text-center">CGS Games Terms of Use</h1>
       <div dangerouslySetInnerHTML={{ __html: rawHTML }}></div>
-      <Footer />
     </main>
   );
 }
