+++ components/gamesDeck.tsx
@@ -1,15 +1,22 @@
 import Game from "@/lib/game";
-import { Card, CardBody, CardFooter, CardHeader } from "@nextui-org/card";
+import { Card, CardBody, CardFooter } from "@nextui-org/card";
 import Link from "next/link";
 import Banner from "./banner";
 
 export default function GamesDeck({ games }: { games: Game[] }) {
+  const isEmpty = games.length == 0;
+  if (isEmpty) {
+    return <h3 className="text-center">No games found.</h3>;
+  }
   return (
     <div className="max-w-[800px] gap-2 grid grid-cols-1">
       {games.map((game, index) => {
         return (
-          <Card key={index} className="border-none bg-slate-400">
-            <CardBody className="overflow-visible p-0">
+          <Card
+            key={index}
+            className="border-none bg-slate-400 flex items-center"
+          >
+            <CardBody>
               <center>
                 <Banner
                   home={`/${game.username}/${game.slug}`}
@@ -19,8 +26,10 @@ export default function GamesDeck({ games }: { games: Game[] }) {
               </center>
             </CardBody>
             <CardFooter className="justify-between text-small text-gray-800">
-              {game.copyright && <p>Copyright of {game.copyright}</p>}
-              <p>
+              {game.copyright && (
+                <p className="ml-4 mr-4">Copyright of {game.copyright}</p>
+              )}
+              <p className="ml-4 mr-4">
                 Uploaded by{" "}
                 <Link href={`/${game.username}`}>{game.username}</Link>
               </p>
