+++ components/usernameForm.tsx
@@ -1,0 +1,134 @@
+"use client";
+
+import { UserContext } from "@/lib/context";
+import { db } from "@/lib/firebase/firebase";
+import { doc, getDoc, writeBatch } from "firebase/firestore";
+import debounce from "lodash.debounce";
+import { useCallback, useContext, useEffect, useState } from "react";
+
+export default function UsernameForm() {
+  const [formValue, setFormValue] = useState("");
+  const [isValid, setIsValid] = useState(false);
+  const [loading, setLoading] = useState(false);
+
+  const { user, username } = useContext(UserContext);
+
+  const onSubmit = async (e: { preventDefault: () => void }) => {
+    e.preventDefault();
+
+    // Create refs for both documents
+    const userDoc = doc(db, `users/${user?.uid}`);
+    const usernameDoc = doc(db, `usernames/${formValue}`);
+
+    // Commit both docs together as a batch write.
+    const batch = writeBatch(db);
+    batch.set(userDoc, {
+      username: formValue,
+    });
+    batch.set(usernameDoc, { uid: user?.uid });
+
+    await batch.commit();
+  };
+
+  const onChange = (e: { target: { value: string } }) => {
+    // Force form value typed in form to match correct format
+    const val = e.target.value.toLowerCase();
+    const re = /^(?=[a-zA-Z0-9._\-]{3,15}$)(?!.*[_.]{2})[^_.].*[^_.]$/;
+
+    // Only set form value if length is < 3 OR it passes regex
+    if (val.length < 3) {
+      setFormValue(val);
+      setLoading(false);
+      setIsValid(false);
+    }
+    if (
+      val === "browse" ||
+      val === "upload" ||
+      val === "terms" ||
+      val === "privacy" ||
+      val === "api"
+    ) {
+      setFormValue(val);
+      setLoading(false);
+      setIsValid(false);
+    }
+
+    if (re.test(val)) {
+      setFormValue(val);
+      setLoading(true);
+      setIsValid(false);
+    }
+  };
+
+  // Hit the database for username match after each debounced change
+  // useCallback is required for debounce to work
+  const checkUsername = useCallback(
+    debounce(async (username) => {
+      if (username.length >= 3) {
+        const ref = doc(db, `usernames/${username}`);
+        const exists = (await getDoc(ref)).exists();
+        console.log(
+          "Firestore read executed! " + username + " exists: " + exists,
+        );
+        setIsValid(!exists);
+        setLoading(false);
+      }
+    }, 500),
+    [],
+  );
+
+  useEffect(() => {
+    checkUsername(formValue);
+  }, [formValue, checkUsername]);
+
+  return (
+    !username && (
+      <section>
+        <h2>Choose Username</h2>
+        <form onSubmit={onSubmit}>
+          <input
+            name="username"
+            placeholder="myname"
+            value={formValue}
+            onChange={onChange}
+          />
+          <UsernameMessage
+            username={formValue}
+            isValid={isValid}
+            loading={loading}
+          />
+          <button type="submit" className="btn-green" disabled={!isValid}>
+            Choose
+          </button>
+          <div>
+            Username: {formValue}
+            <br />
+            Loading: {loading.toString()}
+            <br />
+            Username Valid: {isValid.toString()}
+          </div>
+        </form>
+      </section>
+    )
+  );
+}
+
+function UsernameMessage({
+  username,
+  isValid,
+  loading,
+}: {
+  username: string;
+  isValid: boolean;
+  loading: boolean;
+}) {
+  if (loading) {
+    return <p>Checking...</p>;
+  } else if (isValid) {
+    return <p className="text-success">{username} is available!</p>;
+  } else if (username && !isValid) {
+    return <p className="text-danger">That username is taken!</p>;
+  } else {
+    return <p></p>;
+  }
+}
